update filter component
