+++ app/[locale]/analytics/components/analytics-sidebar-layout.tsx
@@ -12,8 +12,8 @@ import {
 } from '@/config/graphql/analaytics-queries';
 import { GraphQL } from '@/lib/api';
 import { cn } from '@/lib/utils';
-import { SidebarLayout } from './sidebar-layout';
-import { SidebarDefaultLayout } from './SidebarDefaultLayout';
+import { DefaultWindow } from './default-output-window';
+import { OutputWindow } from './output-window';
 import styles from './styles.module.scss';
 
 interface DashboardLayoutProps {
@@ -48,7 +48,7 @@ export function AnalyticsDashboardLayout({ children }: DashboardLayoutProps) {
           )}
         >
           <main className={cn(styles.Main, 'px-4', 'py-6')}>{children}</main>
-          <SidePaneLayout />
+          <OutputWindowComponent />
         </div>
       ) : (
         <div className="flex h-[100vh] flex-col  place-content-center items-center">
@@ -60,7 +60,7 @@ export function AnalyticsDashboardLayout({ children }: DashboardLayoutProps) {
   );
 }
 
-function SidePaneLayout() {
+export function OutputWindowComponent() {
   const searchParams = useSearchParams();
   const indicator = searchParams.get('indicator');
   const time_period = searchParams.get('time-period') || '2023_08';
@@ -119,7 +119,7 @@ function SidePaneLayout() {
     );
   return region !== null && region.length > 0
     ? sidePaneData.isFetched && (
-        <SidebarLayout
+        <OutputWindow
           data={
             sidePaneData?.data[
               boundary === 'district' ? 'districtViewData' : 'revCircleViewData'
@@ -130,7 +130,7 @@ function SidePaneLayout() {
         />
       )
     : sidePaneData.isFetched && (
-        <SidebarDefaultLayout
+        <DefaultWindow
           chartData={
             sidePaneData?.data[
               boundary === 'district' ? 'districtViewData' : 'revCircleViewData'
