+++ package.json
@@ -27,7 +27,7 @@
     "next-intl": "^3.4.0",
     "next-usequerystate": "^1.8.1",
     "nextjs-toploader": "^1.3.2",
-    "opub-ui": "^0.3.21",
+    "opub-ui": "^0.3.29",
     "react": "^18.2.0",
     "react-aria": "3.22.0",
     "react-dom": "^18.2.0",
